--- frontend/src/@state/authState.js
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const loginState = atom({
-  key: 'loginState',
-  default: null,
-});
