+++ frontend/src/pages/Login/Login.jsx
@@ -1,11 +1,8 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
-import { useSetRecoilState } from 'recoil';
-import { loginState } from '@/@state/authState';
-import axios from 'axios';
-import { useNavigate } from 'react-router-dom';
 
 import { Button, Container, Input, Form } from '../../components';
+import { requestLogin } from '@/api/requestLogin';
 
 const Li = styled.li`
   display: flex;
@@ -27,45 +24,17 @@ function Login() {
   const handleEmailChange = (e) => setEmail(e.target.value);
   const handlePasswordChange = (e) => setPassword(e.target.value);
 
-  const navigate = useNavigate();
+  const login = requestLogin();
 
-  // Recoil 상태 설정
-  const setLogin = useSetRecoilState(loginState);
+  const LoginData = {
+    email,
+    loginPassword: password,
+  };
 
-  // 로그인 요청 
   const handleLoginRequest = async (event) => {
     event.preventDefault();
-
-    const LoginData = {
-      email,
-      loginPassword: password,
-    };
-    
-    console.log(LoginData);
-
-    try {
-      const response = await axios.post('/api/login', LoginData);
-      alert('환영합니다 ' + response.data.message + '님!');
-      // 로그인 상태 업데이트
-      setLogin(response.data.user);
-      // 페이지 이동
-      navigate('/');
-
-
-    } catch (error) {
-      if (error.response) {
-        // 요청이 이루어졌으나 서버가 2xx 범위가 아닌 상태 코드로 응답
-        alert('로그인 실패: ' + error.response.data.message);
-      } else if (error.request) {
-        // 요청 실패
-        alert('로그인 요청 실패: 서버에서 응답이 없습니다.');
-      } else {
-        // 요청 설정 중, 문제발생
-        alert('로그인 요청 중 오류 발생: ' + error.message);
-      }
-    }
-  }
-
+    login(LoginData);
+  };
 
   return (
     <Container as="section" align="center" width="widthMedium">
