+++ frontend/src/states/authState.js
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const accessTokenState = atom({
+  key: 'accessTokenState',
+  default: sessionStorage.getItem('accessToken') || '',
+});
